Identify bot PRs when a Jira issue can't be identified.
## Prompt

When a Jira issue is unable to be identified and the user creating 
the PR is a bot, set the Jira issue to "BOT" instead of "UNKNOWN".  
If the user creating the PR is not a bot, use the existing logic and 
set the Jira issue to "UNKNOWN".  Users that are bots are indicated 
by a bot marker "[bot]" in their username.